[wip] Simplify CLI args: commit to preprocess
