--- bin/chat2mark-convert
@@ -1,3 +1,0 @@
-#!/usr/bin/env node
-
-require('../dist/bin/app-convert.js');
